Fixed issues with API
